chore: change service names, application.yml files, build.gradle.kts, add gradle build files in root module
- change service name for auth-service and server
- change settings to eureka-server/client in application.yml files for services
- add gradle build files, root build.gradle.kts and settings.gradle.kts files for root module
