+++ auth-service/.gitattributes
