+++ auth-service/.gitignore
