+++ auth-service/src/test/java/auth/AuthServiceApplicationTests.java
@@ -1,0 +1,15 @@
+package auth;
+
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AuthServiceApplicationTests {
+
+    @Test
+    @Disabled
+    void contextLoads() {
+    }
+
+}
