+++ k8s-server/README.md
@@ -1,0 +1,1 @@
+# message-hub - k8s-server
