+++ k8s-server/build.gradle.kts
@@ -1,0 +1,35 @@
+plugins {
+	java
+	id("org.springframework.boot")
+	id("io.spring.dependency-management")
+}
+
+java {
+	toolchain {
+		languageVersion = JavaLanguageVersion.of(21)
+	}
+}
+
+extra["springCloudVersion"] = "2024.0.0"
+
+dependencies {
+	implementation("org.springframework.boot:spring-boot-starter-actuator")
+	implementation("org.springframework.boot:spring-boot-starter-security")
+	implementation("org.springframework.boot:spring-boot-starter-web")
+
+    implementation("org.springframework.cloud:spring-cloud-starter-netflix-eureka-server")
+
+//    implementation("org.springframework.cloud:spring-cloud-starter-kubernetes")
+//    implementation("org.springframework.cloud:spring-cloud-starter-kubernetes-config")
+//    implementation("org.springframework.cloud:spring-cloud-starter-kubernetes-secrets")
+    runtimeOnly("io.micrometer:micrometer-registry-prometheus")
+	testImplementation("org.springframework.boot:spring-boot-starter-test")
+	testImplementation("org.springframework.security:spring-security-test")
+	testRuntimeOnly("org.junit.platform:junit-platform-launcher")
+}
+
+dependencyManagement {
+	imports {
+		mavenBom("org.springframework.cloud:spring-cloud-dependencies:${property("springCloudVersion")}")
+	}
+}
