+++ k8s-server/src/main/resources/application.yml
@@ -1,0 +1,14 @@
+server:
+  port: 8761
+
+spring:
+  application:
+    name: k8s-server
+eureka:
+  instance:
+    hostname: localhost
+  client:
+    register-with-eureka: false
+    fetch-registry: false
+    serviceUrl:
+      defaultZone: "https://${eureka.instance.hostname}:${server.port}/eureka/"
