+++ k8s-server/src/test/java/server/ServerApplicationTests.java
@@ -1,0 +1,15 @@
+package server;
+
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ServerApplicationTests {
+
+	@Test
+    @Disabled
+	void contextLoads() {
+	}
+
+}
