--- message-hub-auth-service/settings.gradle.kts
@@ -1,1 +1,0 @@
-rootProject.name = "message-hub-auth-service"
