--- message-hub-auth-service/src/main/java/messagehub/auth/MessageHubAuthServiceApplication.java
@@ -1,17 +1,0 @@
-package messagehub.auth;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-
-@EnableFeignClients
-@EnableDiscoveryClient
-@SpringBootApplication
-public class MessageHubAuthServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(MessageHubAuthServiceApplication.class, args);
-    }
-
-}
