--- message-hub-auth-service/src/main/resources/application-dev.yml
@@ -1,5 +1,0 @@
-spring:
-  datasource:
-    url: jdbc:postgresql://localhost:5432/stacover_auth_service
-    username: postgres
-    password: root
