--- message-hub-auth-service/src/test/java/messagehub/auth/MessageHubAuthServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package messagehub.auth;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MessageHubAuthServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
