--- message-hub-eureka-server/README.md
@@ -1,1 +1,0 @@
-# message-hub - Eureka-Server
