--- message-hub-eureka-server/settings.gradle.kts
@@ -1,1 +1,0 @@
-rootProject.name = "message-hub-eureka-server"
