--- message-hub-eureka-server/src/main/java/messagehub/server/MessageHubServerApplication.java
@@ -1,15 +1,0 @@
-package messagehub.server;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
-
-@SpringBootApplication
-@EnableEurekaServer
-public class MessageHubServerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(MessageHubServerApplication.class, args);
-	}
-
-}
