--- message-hub-eureka-server/src/main/resources/application.yml
@@ -1,10 +1,0 @@
-server:
-  port: 8761
-
-spring:
-  application:
-    name: message-hub-server
-  cloud:
-    kubernetes:
-      discovery:
-        enabled: true
