--- message-hub-eureka-server/src/test/java/messagehub/server/MessageHubEurekaServerApplicationTests.java
@@ -1,13 +1,0 @@
-package messagehub.server;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MessageHubEurekaServerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
