+++ settings.gradle.kts
@@ -1,0 +1,2 @@
+rootProject.name = "message-hub"
+include("k8s-server", "auth-service")
